+++ testsrc/sstest.cpp
@@ -1,5 +1,4 @@
 #include "setsketch.h"
-#include "blaze/Math.h"
 #include "hll.h"
 #include <chrono>
